+++ apps/docs/app/status/page.tsx
@@ -1,12 +1,12 @@
 'use client'
 
+import { Link } from 'next-view-transitions'
 import { allComponents } from 'content'
 import {
   GdsBadge,
   GdsDivider,
   GdsFlex,
   GdsGrid,
-  GdsLink,
   GdsText,
 } from '$/import/components'
 import { format, parseISO } from 'date-fns'
@@ -24,13 +24,9 @@ export default function Status() {
         .filter((component) => component._raw.sourceFileName === 'index.mdx')
         .map((component, index) => (
           <GdsGrid columns="3" gap="xl" key={`key-${index}`}>
-            <GdsLink
-              href={component.url_path}
-              title="Component"
-              variant="secondary"
-            >
+            <Link href={component.url_path} title="Component">
               {component.title}
-            </GdsLink>
+            </Link>
             {component.status ? (
               <GdsBadge variant="notice">{component.status}</GdsBadge>
             ) : (